requirements.txt
asgi-redis==1.4.2
asgiref==1.1.2
asn1crypto==0.24.0
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
certifi==2019.9.11
cffi==1.12.2
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
daphne==1.4.2
Django==2.1.4
django-admin-interface==0.5.8
django-colorfield==0.1.15
django-cors-headers==2.1.0
django-djconfig==0.8.0
django-endless-pagination-vue==1.4
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-haystack==2.8.1
django-hitcount==1.2.2
django-infinite-scroll-pagination==0.2.0
django-musette==0.3.2
django-spirit==0.6.2
djangorestframework==3.8.2
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
inflection==0.3.1
mistune==0.7.4
msgpack-python==0.5.6
newsapi-python==0.2.5
olefile==0.45.1
Pillow==5.2.0
praw==6.3.1
prawcore==1.0.1
psycopg2==2.7.3.1
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
python-decouple==3.1
pytz==2018.7
redis==3.0.1
requests==2.21.0
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
uni-slugify==0.1.4
Unidecode==1.0.22
update-checker==0.16
urllib3==1.24.3
websocket-client==0.56.0
Whoosh==2.7.4
zope.interface==4.6.0
